chore(cli/setup): Make setup/ui ESM compatible
