+++ packages/cli/src/commands/setup/ui/helpers/helpers.js
@@ -1,0 +1,9 @@
+export function createHandler(componentName) {
+  return async function handler(argv) {
+    const { handler: importedHandler } = await import(
+      `../libraries/${componentName}Handler.js`
+    )
+
+    return importedHandler(argv)
+  }
+}
